Custom Install Path via Registry
Also fixed the pip installation error - https://stackoverflow.com/questions/31172719/pip-install-access-denied-on-windows
If the registry key doesn't exist, then it will create the key and set it to `%LocalAppData%` (which is the default path as of right now)
`-U` is an alias for `--upgrade` - idk y I did this but I did :\

I have fully tested this. I successfully installed the RLBotGUI onto my micro sd card (mostly bc I also have my programming projects on it, so now I can have the RLBot virtual environment where ever I go now). It also properly installs to the default path.

It would be nice if there was a UI component that modified the registry entry, just to make this feature more user-friendly. (Extra -> Install path?)