Add Support for Custom Maps (#135)
We first find the
CookedPCConsole directory based on whether a user is on
Epic or Steam. Then we swap a real map with the custom
map for the match to start. Once the match is loaded,
we immediately swap the real map file back.

One feature missing is that if a user is on Steam and
has installed the game on a different drive, we can't
detect that.