Use story-mode specific version of blind-and-deaf (#141)
* Use story-mode specific version of blind-and-deaf

Darxeal added an additional bot config that adds a weakness
to blind-and-deaf. This is a bit necessary since a recent
change in blind-and-deaf fixed some quirks which was the only way
to beat blind-and-deaf previously (but also provided for an
inconsistent experience). So hopefully this is a net postive.

* Update display name for BAF to "Flying Panda"

There was discussion about the name in the discord and Will suggested
Flying Panda.