Make repo downloading code more generic and use it for maps
While updating map pack has custom logic, the BotDownloader
is a generic repo downloader. So we rename it to be more
generic and download it.

Also removed the "branch" option since the new botpack updater
strictly is tied to the master branch. So the branch argument is no
longer a real option.

Made the UI modal popup also generic so can be used for the map pack.