Updating and Downloading RLBot MapPack (#144)
* Make repo downloading code more generic and use it for maps

While updating map pack has custom logic, the BotDownloader
is a generic repo downloader. So we rename it to be more
generic and download it.

Also removed the "branch" option since the new botpack updater
strictly is tied to the master branch. So the branch argument is no
longer a real option.

Made the UI modal popup also generic so can be used for the map pack.

* Update settings after downloading map pack

Map pack is added to "bot folders" setting. Note this folder
already supports scanning for bots, scripts and maps. So it
is the right place for this setting.

* Add map pack hydrating and refactor into a class

Hydrating means fetching the actual maps. This commit establishes
the full workflow:

- Check if there is a newer release (using index.json)
- If there is a new release, download the full repo
- Compare old and new index.json and for any updated revision,
fetch the updated map

* Bump release to 0.0.101