Fix AttributeError on Bot Own-goal
If the bot team touches the ball before the packet is grabbed in manage_game_state (so before any bot loads in, so when loading 16 bots for example), the last touch for the player team can be None.

This usually doesn't happen because when the while loop runs before a bot touches the ball, the default last touch has attributes team: 0, player_index: 0, so when the packet is loaded, the player team last touch updates, so it's not None.

Also adds a check for player_name to make sure the player actually touched the ball (player_name is an empty string in the default last touch) for a human goal.

To reproduce, put a time.sleep(10) before the while loop and have a bot own-goal.