Use absolute path for ping command in installer
Some users do not have C:\System32 in their PATH environment variable.
Ping is in C:\System32 so these users are unable to install the required
dependencies using the alternative installer.
All Windows machines are supposed to have ping, so we use an absolute
path to it for users that do not have it in their environment variables.